slimmed down logic
