+++ bopep/design/ma7631si.code-workspace
@@ -1,0 +1,11 @@
+{
+	"folders": [
+		{
+			"path": "../../../.."
+		},
+		{
+			"path": "../../../../../../srv/data1/general"
+		}
+	],
+	"settings": {}
+}
